seperated BoardSidebar out of BoardContainer
